agent/pyproject.toml
[project]
name = "mercury-agent" 
version = "0.1.0"
description = "FactSet terminal backend on A2A Agent Engine"
requires-python = ">=3.11,<3.13"
dependencies = [
    # GCP Infrastructure & ADK
    "google-cloud-aiplatform[agent_engines,adk,langchain]>=1.112.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-genai>=0.2.2",
    "google-adk>=0.1.0",
    # API & Local Server (ADDED THESE)
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    # A2A Protocol SDK
    "a2a-sdk>=0.3.4",
    # Reasoning Engine requirements
    "cloudpickle==3.0.0",
    "nest_asyncio>=1.6.0",
    "pydantic",
    "python-dotenv",
    # FactSet & MCP SDKs
    "fds.sdk.utils>=2.1.0",
    "mcp>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "tools"]
